# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Management Api Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get users",
                "operationId": "2935e090d80a55bdd629d8f62fc7c864",
                "parameters": [
                    {
                        "name": "include",
                        "in": "query",
                        "description": "Load related resources (comma separated)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "company,ownedCompany,projects,assignedProjects,tasks,assignedTasks"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get users successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "milad"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "rezai"
                                                },
                                                "mobile": {
                                                    "type": "string",
                                                    "example": "09167516826"
                                                },
                                                "profile_photo_path": {
                                                    "type": "string",
                                                    "example": "image/image.png"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/users/{user}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user",
                "operationId": "74f2f08e33fae533221a2db8a4eb5805",
                "parameters": [
                    {
                        "name": "include",
                        "in": "query",
                        "description": "Load related resources (comma separated)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "company,ownedCompany,projects,assignedProjects,tasks,assignedTasks"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get users successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "milad"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "rezai"
                                                },
                                                "mobile": {
                                                    "type": "string",
                                                    "example": "09167516826"
                                                },
                                                "profile_photo_path": {
                                                    "type": "string",
                                                    "example": "image/image.png"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "User update",
                "operationId": "be257dbdfdf00f9d727844bd30629fe8",
                "requestBody": {
                    "description": "User update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "mobile": {
                                        "description": "mobile number",
                                        "type": "string",
                                        "maxLength": 11,
                                        "example": "09167516826",
                                        "nullable": true
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "milad",
                                        "nullable": true
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "rezai",
                                        "nullable": true
                                    },
                                    "profile_photo_path": {
                                        "type": "file",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "update project successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "milad"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "rezai"
                                                },
                                                "mobile": {
                                                    "type": "string",
                                                    "example": "09167516826"
                                                },
                                                "profile_photo_path": {
                                                    "type": "string",
                                                    "example": "image/image.png"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "delete a user",
                "operationId": "c1db2101db7a57c130432c374ee463a6",
                "responses": {
                    "204": {
                        "description": "user delete successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}